test: consolidate tax ID tests into string_test.go
Move EIN and SSN test coverage into the main string rule test file and update spell-check configuration.
